# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Pathwise CC API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface AuthUser {
  id: string;
  email?: string | null;
  firstName?: string | null;
  lastName?: string | null;
  profileImageUrl?: string | null;
}

export interface GetCurrentAuthUserResponse {
  user: AuthUser;
}

export interface StudentProfile {
  id: number;
  userId: string;
  fullName?: string | null;
  communityCollege?: string | null;
  currentGpa?: number | null;
  intendedMajor?: string | null;
  careerGoal?: string | null;
  financialSituation?: string | null;
  transferTimeline?: string | null;
  geographicPreference?: string | null;
  targetUniversities?: string[] | null;
  longTermAspirations?: string | null;
  isFirstGen?: string | null;
  interests?: string[] | null;
  completionPercent?: number | null;
  createdAt?: string;
  updatedAt?: string;
}

export interface CreateProfileBody {
  fullName?: string;
  communityCollege?: string;
  currentGpa?: number;
  intendedMajor?: string;
  careerGoal?: string;
  financialSituation?: string;
  transferTimeline?: string;
  geographicPreference?: string;
  targetUniversities?: string[];
  longTermAspirations?: string;
  isFirstGen?: string;
  interests?: string[];
}

export interface UpdateProfileBody {
  fullName?: string;
  communityCollege?: string;
  currentGpa?: number;
  intendedMajor?: string;
  careerGoal?: string;
  financialSituation?: string;
  transferTimeline?: string;
  geographicPreference?: string;
  targetUniversities?: string[];
  longTermAspirations?: string;
  isFirstGen?: string;
  interests?: string[];
  completionPercent?: number;
}

export interface Course {
  id: number;
  profileId: number;
  courseCode?: string | null;
  courseName: string;
  units?: number | null;
  grade?: string | null;
  status?: string | null;
  term?: string | null;
  createdAt?: string;
}

export interface CreateCourseBody {
  courseCode?: string;
  courseName: string;
  units?: number;
  grade?: string;
  status?: string;
  term?: string;
}

export interface UpdateCourseBody {
  courseCode?: string;
  courseName?: string;
  units?: number;
  grade?: string;
  status?: string;
  term?: string;
}

export interface GpaSummary {
  estimatedGpa: number;
  totalUnits: number;
  completedUnits: number;
  inProgressUnits: number;
  courseCount: number;
}

export interface ScoreBreakdown {
  majorAlignment: number;
  gpaFit: number;
  coursework?: number;
  costFit: number;
  locationFit: number;
  careerFit: number;
  experienceAccess?: number;
  timeFit?: number;
  supportFit?: number;
}

export interface UniversityMatch {
  universityId: string;
  name: string;
  system?: string;
  location?: string;
  compatibilityScore: number;
  fitLabel: string;
  fitColor?: string;
  gpaRangeMin?: number;
  gpaRangeRecommended?: number;
  costCategory?: string;
  transferFriendliness?: number;
  officialTransferUrl?: string;
  notes?: string;
  breakdown?: ScoreBreakdown;
  majors?: string[];
  careerTags?: string[];
}

export interface PathwayReport {
  type: string;
  university: string;
  compatibilityScore: number;
  whyItFits?: string;
  concerns?: string;
  gpaTarget?: number;
  courseGaps?: string[];
  transferTimeline?: string;
  scholarshipOptions?: string[];
  internshipRecommendations?: string[];
  extracurricularRecommendations?: string[];
  risks?: string[];
  nextSteps?: string[];
}

export interface Pathway {
  id: number;
  profileId: number;
  universityId?: string | null;
  compatibilityScore?: number | null;
  pathwayType?: string | null;
  reportJson?: PathwayReport;
  isSelected?: string | null;
  createdAt?: string;
}

export interface Guidebook {
  id: number;
  pathwayId: number;
  profileId: number;
  contentMarkdown?: string | null;
  title?: string | null;
  createdAt?: string;
  updatedAt?: string;
}

export interface Scholarship {
  id: string;
  name: string;
  amount: string;
  deadline?: string;
  minGpa?: number | null;
  eligibilityTags?: string[];
  description?: string;
  sourceUrl?: string;
  sourceDate?: string;
  isEstimate?: boolean;
}

export interface Opportunity {
  id: string;
  name: string;
  type: string;
  description?: string;
  eligibilityTags?: string[];
  applicationUrl?: string;
  sourceDate?: string;
  isEstimate?: boolean;
}

export interface DashboardSummary {
  profileCompletionPercent: number;
  totalCourses: number;
  completedCourses: number;
  inProgressCourses: number;
  estimatedGpa?: number | null;
  savedPathwaysCount: number;
  guidebooksCount: number;
  topMatchUniversity?: string | null;
  topMatchScore?: number | null;
  nextActions: string[];
}