# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Pathwise CC API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/user:
    get:
      operationId: getCurrentAuthUser
      tags: [auth]
      summary: Get current authenticated user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GetCurrentAuthUserResponse"
        "401":
          description: Not authenticated
...
  /profiles/{profileId}/courses:
    get:
      operationId: getCourses
      tags: [courses]
      summary: List courses for profile
      parameters:
        - name: profileId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of courses
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Course"
    post:
      operationId: addCourse
      tags: [courses]
      summary: Add course to profile
      parameters:
        - name: profileId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCourseBody"
      responses:
        "201":
          description: Course added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Course"
  /courses/{courseId}:
    patch:
      operationId: updateCourse
      tags: [courses]
      summary: Update course
      parameters:
        - name: courseId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateCourseBody"
      responses:
        "200":
          description: Updated course
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Course"
    delete:
      operationId: deleteCourse
      tags: [courses]
      summary: Delete course
      parameters:
        - name: courseId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /profiles/{profileId}/gpa-summary:
    get:
      operationId: getGpaSummary
      tags: [courses]
      summary: Get GPA summary for profile
      parameters:
        - name: profileId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: GPA summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GpaSummary"
...
  /dashboard-summary/{profileId}:
    get:
      operationId: getDashboardSummary
      tags: [profiles]
      summary: Get dashboard summary for profile
      parameters:
        - name: profileId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Dashboard summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardSummary"
components:
  schemas:
...
    GetCurrentAuthUserResponse:
      type: object
      properties:
        user:
          $ref: "#/components/schemas/AuthUser"
      required: [user]
    AuthUser:
      type: object
      properties:
        id:
          type: string
        email:
          type: string
          nullable: true
        firstName:
          type: string
          nullable: true
        lastName:
          type: string
          nullable: true
        profileImageUrl:
          type: string
          nullable: true
      required: [id]
...
    Course:
      type: object
      properties:
        id:
          type: integer
        profileId:
          type: integer
        courseCode:
          type: string
          nullable: true
        courseName:
          type: string
        units:
          type: number
          nullable: true
        grade:
          type: string
          nullable: true
        status:
          type: string
          nullable: true
        term:
          type: string
          nullable: true
        createdAt:
          type: string
      required: [id, profileId, courseName]
    CreateCourseBody:
      type: object
      properties:
        courseCode:
          type: string
        courseName:
          type: string
        units:
          type: number
        grade:
          type: string
        status:
          type: string
        term:
          type: string
      required: [courseName]
    UpdateCourseBody:
      type: object
      properties:
        courseCode:
          type: string
        courseName:
          type: string
        units:
          type: number
        grade:
          type: string
        status:
          type: string
        term:
          type: string
    GpaSummary:
      type: object
      properties:
        estimatedGpa:
          type: number
        totalUnits:
          type: number
        completedUnits:
          type: number
        inProgressUnits:
          type: number
        courseCount:
          type: integer
      required: [estimatedGpa, totalUnits, completedUnits, inProgressUnits, courseCount]
...
    DashboardSummary:
      type: object
      properties:
        profileCompletionPercent:
          type: number
        totalCourses:
          type: integer
        completedCourses:
          type: integer
        inProgressCourses:
          type: integer
        estimatedGpa:
          type: number
          nullable: true
        savedPathwaysCount:
          type: integer
        guidebooksCount:
          type: integer
        topMatchUniversity:
          type: string
          nullable: true
        topMatchScore:
          type: integer
          nullable: true
        nextActions:
          type: array
          items:
            type: string
      required: [profileCompletionPercent, totalCourses, completedCourses, inProgressCourses, savedPathwaysCount, guidebooksCount, nextActions]