# lib/api-zod/src/generated/types/course.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Pathwise CC API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Course {
  id: number;
  profileId: number;
  courseCode?: string | null;
  courseName: string;
  units?: number | null;
  grade?: string | null;
  status?: string | null;
  term?: string | null;
  createdAt?: string;
}