# lib/api-zod/src/generated/types/dashboardSummary.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Pathwise CC API specification
 * OpenAPI spec version: 0.1.0
 */

export interface DashboardSummary {
  profileCompletionPercent: number;
  totalCourses: number;
  completedCourses: number;
  inProgressCourses: number;
  estimatedGpa?: number | null;
  savedPathwaysCount: number;
  guidebooksCount: number;
  topMatchUniversity?: string | null;
  topMatchScore?: number | null;
  nextActions: string[];
}