# lib/api-zod/src/generated/types/getCurrentAuthUserResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Pathwise CC API specification
 * OpenAPI spec version: 0.1.0
 */
import type { AuthUser } from "./authUser";

export interface GetCurrentAuthUserResponse {
  user: AuthUser;
}