# lib/api-zod/src/generated/types/gpaSummary.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Pathwise CC API specification
 * OpenAPI spec version: 0.1.0
 */

export interface GpaSummary {
  estimatedGpa: number;
  totalUnits: number;
  completedUnits: number;
  inProgressUnits: number;
  courseCount: number;
}