# lib/api-zod/src/generated/types/updateCourseBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Pathwise CC API specification
 * OpenAPI spec version: 0.1.0
 */

export interface UpdateCourseBody {
  courseCode?: string;
  courseName?: string;
  units?: number;
  grade?: string;
  status?: string;
  term?: string;
}